Refactor flaky tests
