+++ stake-pool/program/tests/initialize.rs
@@ -1273,23 +1273,3 @@ async fn fail_with_bad_reserve() {
         );
     }
 }
-
-#[tokio::test]
-async fn success_with_deposit_authority() {
-    let (mut banks_client, payer, recent_blockhash) = program_test().start().await;
-    let deposit_authority = Keypair::new();
-    let stake_pool_accounts = StakePoolAccounts::new_with_deposit_authority(deposit_authority);
-    let deposit_authority = stake_pool_accounts.stake_deposit_authority;
-    stake_pool_accounts
-        .initialize_stake_pool(&mut banks_client, &payer, &recent_blockhash, 1)
-        .await
-        .unwrap();
-
-    // Stake pool now exists
-    let stake_pool_account =
-        get_account(&mut banks_client, &stake_pool_accounts.stake_pool.pubkey()).await;
-    let stake_pool =
-        try_from_slice_unchecked::<state::StakePool>(stake_pool_account.data.as_slice()).unwrap();
-    assert_eq!(stake_pool.stake_deposit_authority, deposit_authority);
-    assert_eq!(stake_pool.sol_deposit_authority.unwrap(), deposit_authority);
-}
